backend/requirements.txt
--only-binary :all:

fastapi
uvicorn
python-dotenv
httpx
pydantic
openai